+++ src/robothub_depthai/components/stereo.py
@@ -1,4 +1,4 @@
-from robothub_depthai.components import Streamable
+from robothub_depthai.components.streamable import Streamable
 
 
 class Stereo(Streamable):
